requirements.txt
google-generativeai>=0.3.0
agent-framework-core>=1.0.0b251114

